before huge rewrite

+++ inc/filler.h
@@ -6,7 +6,7 @@
 /*   By: aburdeni <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/17 18:49:01 by aburdeni          #+#    #+#             */
-/*   Updated: 2018/11/29 22:50:49 by aburdeni         ###   ########.fr       */
+/*   Updated: 2018/11/29 23:25:16 by aburdeni         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,29 +26,29 @@
 typedef struct	s_f
 {
 	char		**board;
-	size_t		n;
-	size_t		x;
+	int			n;
+	int			x;
 
 	char		**token;
-	size_t		tn;
-	size_t		tx;
+	int			tn;
+	int			tx;
 
 	char 		player;
 	char 		enemy;
 
-	size_t		out_n;
-	size_t		out_x;
+	int		out_n;
+	int		out_x;
 }				t_f;
 
 typedef struct	s_out
 {
-	size_t		n;
-	size_t		x;
-	size_t		steps;
-	size_t		t_n;
-	size_t		t_x;
-	size_t		t_n_end;
-	size_t		t_x_end;
+	int		n;
+	int		x;
+	int		steps;
+	int		t_n;
+	int		t_x;
+	int		t_n_end;
+	int		t_x_end;
 }				t_out;
 
 extern t_f		g_f;
@@ -56,7 +56,7 @@ extern t_out	g_out;
 
 void			search_place();
 void			set_token_coord();
-int				valid_position(size_t n, size_t x);
+int				valid_position(int n, int x);
 int				try_put_around(int n, int x, int nDist, int xDist);
 int				try_put_in_token_distance(int n, int x, int nDist, int xDist);
